pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "coinbase-advanced-py>=1.8.2",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]
